Many updates and improvements

+++ part 4/blog-list-app-backend/controllers/users.js
@@ -3,6 +3,7 @@ const userRouter = require('express').Router()
 const User = require('../models/user')
 const Blog = require('../models/blog')
 const bcrypt = require('bcrypt')
+const middleware = require("../utils/middleware")
 
 userRouter.get('/', async (req, res) => {
   const ourUsers = await User.find({}).populate('blogs', { title: 1, author: 1, url: 1, likes: 1 })
@@ -22,16 +23,34 @@ userRouter.get('/:id', async (req, res, next) => {
   } catch(err) {
     next(err)
   }
-
 })
 
+// retrieve a user's liked blogposts
+userRouter.get('/:id/likes', async (req, res, next) => {
+  try {
+
+    const id = req.params.id
+
+    const user = await User.findOne({ username: id })
+    console.log('Trying to view bloglist for user: ', user.username)
+    console.log("User's liked blogposts: ", user.likedBlogs)
+
+    res.status(200).json(user.likedBlogs)
+
+  } catch(err) {
+    next(err)
+  }
+})
+
+// retrieve blogs belonging to a specific user
 userRouter.get('/:id/blogs', async (req, res, next) => {
   try {
 
     const id = req.params.id
 
     const user = await User.findOne({ username: id })
-    console.log('Trying to view bloglist for user: ', user)
+    console.log('Trying to view bloglist for user: ', user.username)
+    console.log("User's liked blogposts: ", user.likedBlogs)
 
     const userBlogsToDisplay = await Blog.find({ userId: user._id })
 
@@ -42,44 +61,53 @@ userRouter.get('/:id/blogs', async (req, res, next) => {
   }
 })
 
-userRouter.delete('/:id', async (req, res, next) => {
+// deletes an entire user profile, and all of the user's associated blogs
+userRouter.delete('/:id', middleware.userExtractor, async (req, res, next) => {
   const id = req.params.id
+  const user = req.user
+  console.log("User resolved from token: ", user)
 
-  try {
-    console.log('request to delete user:', id)
-    const userToDelete = await User.findById(id)
-    console.log('Trying to delete user: ', userToDelete)
+  // if user is authenticated and is trying to delete his own content
+  if(user && (user.id === id)) {
+    try {
+      console.log('request to delete user:', id)
+      const userToDelete = user
+      console.log('Trying to delete user: ', userToDelete.id)
 
-    const userBlogsToRemove = await Blog.find({ userId: userToDelete._id })
-    console.log('User blogs which also need to be removed: ', userBlogsToRemove)
+      // NEED TO CHECK IF THIS IS REDUNDANT
+      const userBlogsToRemove = user
+      console.log('User blogs which also need to be removed: ', userBlogsToRemove)
 
-    if(userBlogsToRemove.length === 0) {
-      console.log('User blog array is empty. Delete the user.')
+      if(userBlogsToRemove.length === 0) {
+        console.log('User blog array is empty. Delete the user.')
+        await User.findByIdAndRemove(userToDelete._id)
+        res.status(204).end()
+      }
+
+      await Blog.deleteMany({ userId: userToDelete._id })
+      console.log('Removed user blogs')
+
       await User.findByIdAndRemove(userToDelete._id)
+      console.log('User removed')
+
       res.status(204).end()
+    } catch(err) {
+      next(err)
     }
-
-    await Blog.deleteMany({ userId: userToDelete._id })
-    console.log('Removed user blogs')
-
-    await User.findByIdAndRemove(userToDelete._id)
-    console.log('User removed')
-
-    res.status(204).end()
-  } catch(err) {
-    next(err)
   }
 
-})
+  res.status(400).json({ error: "You must be logged in to perform this action." })
 
+})
+// sign-up new user route
 userRouter.post('/', async (req, res, next) => {
   try {
     const { username, name, password } = req.body
     console.log('New user attempting to register: ', username, name, password)
 
-    if(password.length <= 3) {
+    if(password.length <= 10) {
       console.log("Registration failed.")
-      return res.status(400).json({ error: 'Password must be longer than 3 characters.' })
+      return res.status(400).json({ error: 'Password must be longer than 10 characters.' })
     }
 
     const saltRounds = 10
